commit from git

+++ DKPlugins/DKCef/DKCefV8.cpp
@@ -394,7 +394,7 @@ bool DKCefV8::PhysicalMemoryUsed(CefArgs args, CefReturn retval){
 
 bool DKCefV8::PhysicalMemoryUsedByApp(CefArgs args, CefReturn retval){
 	DKDEBUGFUNC(args, retval);
-	unsigned int physicalMemory;
+	unsigned long long physicalMemory;
 	if(!DKUtil::PhysicalMemoryUsedByApp(physicalMemory)){ return false; }
 	if(!retval->SetInt(0, (int)physicalMemory)){ return false; }
 	return true;
@@ -541,7 +541,7 @@ bool DKCefV8::VirtualMemoryUsed(CefArgs args, CefReturn retval){
 
 bool DKCefV8::VirtualMemoryUsedByApp(CefArgs args, CefReturn retval){
 	DKDEBUGFUNC(args, retval);
-	unsigned int virtualMemory;
+	unsigned long long virtualMemory;
 	if(!DKUtil::VirtualMemoryUsedByApp(virtualMemory)){ return false; }
 	if(!retval->SetInt(0, (int)virtualMemory)){ return false; }
 	return true;
